fix(expr): remove trivial u64-as-u64 cast from unsigned_abs()
unsigned_abs() on i64 already returns u64; the trailing `as u64` was
flagged by clippy::trivially_copy_pass_by_ref on the newer stable.

+++ crates/expr/src/providers.rs
@@ -150,7 +150,7 @@ impl DiskProvider {
             let result = unsafe { libc::statfs(c_path.as_ptr(), &raw mut stat) };
             if result == 0 {
                 #[cfg(target_os = "linux")]
-                let block_size = stat.f_bsize.unsigned_abs() as u64;
+                let block_size = stat.f_bsize.unsigned_abs();
                 #[cfg(not(target_os = "linux"))]
                 let block_size = u64::from(stat.f_bsize);
                 DiskContext {
